// src/entities/payment/payment.types.ts
export type PaymentInfo = {
	notification_type: string;
	bill_id: string;
	amount: string;
	codepro: string;
	withdraw_amount: string;
	unaccepted: string;
	label: string;
	datetime: string;
	sender: string;
	sha1_hash: string;
	operation_label: string;
	operation_id: string;
	currency: string;
};